在驱动内单独定义出IO

+++ BMS/application/CAN_App.c
@@ -56,7 +56,7 @@ CAN_MsgType CAN_NodeState =
 
 void CAN_SendDemo(void)
 {
-    static int CAN_Time = 0;
+    static uint8 CAN_Time = 0;
     CAN_Time++;
     if (CAN_Time == 1)
     {
